requirements.txt
python-dotenv
sqlalchemy
flask
flask-login
emoji
gunicorn
waitress
